requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.3.0
django-graphql-jwt==0.3.1
graphene==2.1.8
graphene-django==2.10.1
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
idna==2.9
promise==2.3
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
requests==2.23.0
Rx==1.6.1
selenium==3.141.0
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
